Fall back to the config version where there is no native one
expo-application's nativeApplicationVersion is null on the web, so the
settings footer read "LangX 0.0.0" on app.langx.io and every web request
reported the same. Read the version from app.config.ts when the binary
has none, and keep one helper for the footer and the request headers
instead of two copies of the same fallback.

The unknown case no longer sends "0.0.0" either: the server never
forces an update on a missing header, but the placeholder parses and
sorts below every minimum, so it would have locked such a client out.
